build error removed
Former-commit-id: 716290700af5e6dcba22619d12ac39ac52a478b6 [formerly 4e4c2e6e29d45ad74ddbf709af79a7c94fbe2564] [formerly 7bac428b1ffe87f2cec3f210c3b1a90946d00a75 [formerly 70ddf50555ef2cb329077535212d4175e7b352cf]]
Former-commit-id: 78401987ae5069ed108e3f24b1b0b3bfb78a68ad [formerly 6dd0bd6ba1c2de7dad2cd6f98d9fa2dc3d007c3b]
Former-commit-id: b3089b1b9cc1ab255db94949b5797b89a2163eb2

+++ src/node/n_float.cc
@@ -36,9 +36,9 @@ namespace FT{
     void NodeFloat<T>::evaluate(const Data& data, Stacks& stack)
     {
         if(arity['b'])
-            stack.f.row(stack.idx['f']) = stack.b.row(stack.idx['b']).cast<double>();
+            stack.f.row(stack.idx['f']) = stack.b.row(stack.idx['b']).cast<float>();
         else
-            stack.f.row(stack.idx['f']) = stack.c.row(stack.idx['c']).cast<double>();
+            stack.f.row(stack.idx['f']) = stack.c.row(stack.idx['c']).cast<float>();
     }
 #endif
